+++ io/applications/CMakeLists.txt
@@ -38,7 +38,7 @@ if( NOT WIN32 )
     set_target_properties( io-ls PROPERTIES LINK_FLAGS_RELEASE -s )
     install( TARGETS io-ls RUNTIME DESTINATION ${comma_INSTALL_BIN_DIR} COMPONENT Runtime )
     
-    add_executable( io-publish ${dir}/io-publish.cpp )
+    add_executable( io-publish ${dir}/io-publish.cpp ${dir}/detail/publish.h ${dir}/detail/publish.cpp )
     target_link_libraries( io-publish comma_base comma_io comma_application comma_xpath comma_name_value ) # profiler )
     set_target_properties( io-publish PROPERTIES LINK_FLAGS_RELEASE -s )
     install( TARGETS io-publish RUNTIME DESTINATION ${comma_INSTALL_BIN_DIR} COMPONENT Runtime )
@@ -48,8 +48,8 @@ if( NOT WIN32 )
     set_target_properties( io-tee PROPERTIES LINK_FLAGS_RELEASE -s )
     install( TARGETS io-tee RUNTIME DESTINATION ${comma_INSTALL_BIN_DIR} COMPONENT Runtime )
     
-    add_executable( udp-client ${dir}/udp-client.cpp )
-    target_link_libraries( udp-client comma_application comma_io comma_csv comma_base ) # profiler )
+    add_executable( udp-client ${dir}/udp-client.cpp ${dir}/detail/publish.h ${dir}/detail/publish.cpp )
+    target_link_libraries( udp-client comma_application comma_name_value comma_io comma_csv comma_base ) # profiler )
     set_target_properties( udp-client PROPERTIES LINK_FLAGS_RELEASE -s )
     install( TARGETS udp-client RUNTIME DESTINATION ${comma_INSTALL_BIN_DIR} COMPONENT Runtime )
